create message component
